lab 7 is almost complete

+++ server/src/dimka/blinb/collection/objects/Location2D.java
@@ -2,6 +2,6 @@ package dimka.blinb.collection.objects;
 
 public class Location2D extends Location {
     Location2D(String name, Float x, Integer y){
-        super(name, x, y);
+        super(name, x, (long) y);
     }
 }
